requirements.txt
numpy~=1.21.4
keras~=2.7.0
Pillow~=8.4.0
pandas~=1.3.4
Flask~=2.0.2
passlib~=1.7.4
WTForms~=3.0.0
Flask-Login~=0.5.0
Flask-WTF~=1.0.0
tensorflow~=2.7.0